+++ src/components/Markdown/Pages.jsx
@@ -11,11 +11,12 @@ import {
 } from '@material-ui/core';
 import withWidth, { isWidthDown } from '@material-ui/core/withWidth';
 import { Help, School, Menu } from '@material-ui/icons';
-import { Redirect, Link, useLocation, useHistory } from 'react-router-dom';
+import { Redirect, useLocation, useHistory } from 'react-router-dom';
 import clsx from 'clsx';
 import MarkdownRenderer from './Renderer';
 import WhoAmIButton from '../Interfaces/interfaceHelpers/WhoAmIButton';
 import HelpModal from '../Interfaces/interfaceHelpers/HelpModal';
+import TeachersModal from '../Interfaces/interfaceHelpers/TeacherHelpModal';
 import NavDrawer from '../UI/NavDrawer';
 import NavButtons from '../UI/NavButtons';
 import { getFilteredLiveCheckOffsData, getLiveTutorialSelection } from '../../hooks/pages';
@@ -81,6 +82,7 @@ const MarkdownPages = ({
   const [loading, setLoading] = useState(false);
   const [showMenu, setShowMenu] = useState(false);
   const [showHelp, setShowHelp] = useState(false);
+  const [showTeachers, setShowTeachers] = useState(false);
   const [page, setPage] = useState(homePage);
   const [child, setChild] = useState(whoAmI);
   const checkOffs = getFilteredLiveCheckOffsData(a =>
@@ -158,9 +160,9 @@ const MarkdownPages = ({
         wrap: true,
         content: (
           <>
-            <Link to="/teachers" className={classes.navLink}>
+            <button onClick={() => setShowTeachers(true)} className={classes.navLink}>
               Teachers
-            </Link>
+            </button>
             <button onClick={() => setShowHelp(true)} className={classes.navLink}>
               Help!
             </button>
@@ -171,13 +173,13 @@ const MarkdownPages = ({
         ),
         wrappedContent: (
           <List className={classes.linksList}>
-            <ListItem divider button onClick={() => history.push('/teachers')}>
+            <ListItem divider button onClick={() => setShowTeachers(true)}>
               <ListItemIcon>
                 <School />
               </ListItemIcon>
               <ListItemText primary="Teachers" />
             </ListItem>
-            <ListItem divider={child !== null} button onClick={() => history.push('/help')}>
+            <ListItem divider={child !== null} button onClick={() => setShowHelp(true)}>
               <ListItemIcon>
                 <Help />
               </ListItemIcon>
@@ -265,6 +267,7 @@ const MarkdownPages = ({
         whiteList={whieListedPages}
       />
       <HelpModal open={showHelp} onClose={() => setShowHelp(false)} />
+      <TeachersModal open={showTeachers} onClose={() => setShowTeachers(false)} />
     </div>
   );
 };
